refactor: refactored ToolGroups component, separated UI from logic, removed duplicated code and left TODOs
